Fixed auto dismiss notification time default value

+++ app/src/main/java/com/phantom/onetapvideodownload/IpcService.java
@@ -100,7 +100,7 @@ public class IpcService extends IntentService {
 
         final int id = possibleId;
         notificationmanager.notify(id, mBuilder.build());
-        long delayInSeconds = CheckPreferences.notificationDismissTime(this);
+        int delayInSeconds = CheckPreferences.notificationDismissTime(this);
         mHandler.postDelayed(new Runnable() {
             public void run() {
                 notificationmanager.cancel(id);
